Move image services to their own folders
